[feat] Upgrade to webpack5
This is based on my own webpack config which has minor changes.
Required changes:

Added npm dependencies:

  - webpack-cli

Removed npm dependencies:

  - style-loader: Not compatible with the new css extractor

Changed npm dependencies:

  - babel-core -> @babel/core
  - node-sass -> sass  # node-sass is deprecated
  - extract-text-webpack-plugin -> mini-css-extract-plugin

In the webpack config, I was able to remove the require('webpack') line.
The modules pulled from webpack are now included by default.
The valid parameters for devtool changed.
It is now a requirement to state, if one is running webpack in
production mode or not.
The module syntax has been changed.
I updated the plugins in the end of the file. The removed plugins aren't
required any more.

I will need to test the changes with the Lektor plugin.
js code worked for me, I have to validate that css still works.